# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface AuthUser {
  id: string;
  email: string | null;
  firstName: string | null;
  lastName: string | null;
  profileImageUrl: string | null;
}

export interface GetCurrentAuthUserResponse {
  user: AuthUser | null;
}

export interface ExchangeMobileAuthorizationCodeBody {
  code: string;
  code_verifier: string;
  redirect_uri: string;
  state: string;
  nonce?: string | null;
}

export interface ExchangeMobileAuthorizationCodeResponse {
  token: string;
}

export interface LogoutMobileSessionResponse {
  success: boolean;
}

export interface ErrorResponse {
  error: string;
  code?: string | null;
}

export interface Group {
  id: string;
  name: string;
  type: string;
  role: string;
  createdAt: string;
  updatedAt: string;
}

export interface GroupResponse {
  group: Group;
}

export interface ListGroupsResponse {
  groups: Group[];
}

export type CreateGroupBodyType =
  (typeof CreateGroupBodyType)[keyof typeof CreateGroupBodyType];

export const CreateGroupBodyType = {
  family: "family",
  business: "business",
} as const;

export interface CreateGroupBody {
  name: string;
  type: CreateGroupBodyType;
}

export type UpdateGroupBodyType =
  (typeof UpdateGroupBodyType)[keyof typeof UpdateGroupBodyType];

export const UpdateGroupBodyType = {
  family: "family",
  business: "business",
} as const;

export interface UpdateGroupBody {
  name?: string;
  type?: UpdateGroupBodyType;
}

export interface Location {
  id: string;
  groupId: string;
  name: string;
  parentId: string | null;
  qrCode?: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface LocationResponse {
  location: Location;
}

export interface ListLocationsResponse {
  locations: Location[];
}

export interface CreateLocationBody {
  name: string;
  parentId?: string | null;
}

export interface UpdateLocationBody {
  name?: string;
  parentId?: string | null;
  qrCode?: string | null;
}

export interface Bin {
  id: string;
  groupId: string;
  locationId: string;
  name: string;
  qrCode?: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface BinResponse {
  bin: Bin;
}

export interface ListBinsResponse {
  bins: Bin[];
}

export interface CreateBinBody {
  name: string;
  locationId: string;
  qrCode?: string | null;
}

export interface UpdateBinBody {
  name?: string;
  locationId?: string;
  qrCode?: string | null;
}

export interface Item {
  id: string;
  groupId: string;
  binId: string;
  name: string;
  photo?: string | null;
  qrCode?: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface ItemResponse {
  item: Item;
}

export interface ListItemsResponse {
  items: Item[];
}

export interface CreateItemBody {
  name: string;
  binId: string;
  photo?: string | null;
  qrCode?: string | null;
}

export interface UpdateItemBody {
  name?: string;
  binId?: string;
  photo?: string | null;
  qrCode?: string | null;
}

export interface RequestUploadUrlBody {
  name: string;
  size: number;
  contentType: string;
}

export type RequestUploadUrlResponseMetadata = {
  name: string;
  size: number;
  contentType: string;
};

export interface RequestUploadUrlResponse {
  uploadURL: string;
  objectPath: string;
  metadata: RequestUploadUrlResponseMetadata;
}

export type QrLookupResponseType =
  (typeof QrLookupResponseType)[keyof typeof QrLookupResponseType];

export const QrLookupResponseType = {
  bin: "bin",
  location: "location",
  item: "item",
} as const;

export interface QrLookupResponse {
  type: QrLookupResponseType;
  id: string;
  groupId: string;
}

export type ListItemsParams = {
  /**
   * Search query
   */
  q?: string;
  /**
   * Filter by bin
   */
  binId?: string;
};