# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary Get current authenticated user
 */
export const GetCurrentAuthUserResponse = zod.object({
  user: zod.union([
    zod.object({
      id: zod.string(),
      email: zod.string().nullable(),
      firstName: zod.string().nullable(),
      lastName: zod.string().nullable(),
      profileImageUrl: zod.string().nullable(),
    }),
    zod.null(),
  ]),
});

/**
 * @summary Exchange PKCE authorization code for session token (mobile)
 */
export const ExchangeMobileAuthorizationCodeBody = zod.object({
  code: zod.string(),
  code_verifier: zod.string(),
  redirect_uri: zod.string(),
  state: zod.string(),
  nonce: zod.string().nullish(),
});

export const ExchangeMobileAuthorizationCodeResponse = zod.object({
  token: zod.string(),
});

/**
 * @summary Logout mobile session
 */
export const LogoutMobileSessionResponse = zod.object({
  success: zod.boolean(),
});

/**
 * @summary List groups for the current user
 */
export const ListGroupsResponse = zod.object({
  groups: zod.array(
    zod.object({
      id: zod.string(),
      name: zod.string(),
      type: zod.string(),
      role: zod.string(),
      createdAt: zod.coerce.date(),
      updatedAt: zod.coerce.date(),
    }),
  ),
});

/**
 * @summary Create a new group
 */
export const CreateGroupBody = zod.object({
  name: zod.string(),
  type: zod.enum(["family", "business"]),
});

/**
 * @summary Get a group by ID
 */
export const GetGroupParams = zod.object({
  groupId: zod.coerce.string(),
});

export const GetGroupResponse = zod.object({
  group: zod.object({
    id: zod.string(),
    name: zod.string(),
    type: zod.string(),
    role: zod.string(),
    createdAt: zod.coerce.date(),
    updatedAt: zod.coerce.date(),
  }),
});

/**
 * @summary Update a group
 */
export const UpdateGroupParams = zod.object({
  groupId: zod.coerce.string(),
});

export const UpdateGroupBody = zod.object({
  name: zod.string().optional(),
  type: zod.enum(["family", "business"]).optional(),
});

export const UpdateGroupResponse = zod.object({
  group: zod.object({
    id: zod.string(),
    name: zod.string(),
    type: zod.string(),
    role: zod.string(),
    createdAt: zod.coerce.date(),
    updatedAt: zod.coerce.date(),
  }),
});

/**
 * @summary Delete a group
 */
export const DeleteGroupParams = zod.object({
  groupId: zod.coerce.string(),
});

/**
 * @summary List locations in a group
 */
export const ListLocationsParams = zod.object({
  groupId: zod.coerce.string(),
});

export const ListLocationsResponse = zod.object({
  locations: zod.array(
    zod.object({
      id: zod.string(),
      groupId: zod.string(),
      name: zod.string(),
      parentId: zod.string().nullable(),
      qrCode: zod.string().nullish(),
      createdAt: zod.coerce.date(),
      updatedAt: zod.coerce.date(),
    }),
  ),
});

/**
 * @summary Create a location
 */
export const CreateLocationParams = zod.object({
  groupId: zod.coerce.string(),
});

export const CreateLocationBody = zod.object({
  name: zod.string(),
  parentId: zod.string().nullish(),
});

/**
 * @summary Update a location
 */
export const UpdateLocationParams = zod.object({
  groupId: zod.coerce.string(),
  locationId: zod.coerce.string(),
});

export const UpdateLocationBody = zod.object({
  name: zod.string().optional(),
  parentId: zod.string().nullish(),
  qrCode: zod.string().nullish(),
});

export const UpdateLocationResponse = zod.object({
  location: zod.object({
    id: zod.string(),
    groupId: zod.string(),
    name: zod.string(),
    parentId: zod.string().nullable(),
    qrCode: zod.string().nullish(),
    createdAt: zod.coerce.date(),
    updatedAt: zod.coerce.date(),
  }),
});

/**
 * @summary Delete a location
 */
export const DeleteLocationParams = zod.object({
  groupId: zod.coerce.string(),
  locationId: zod.coerce.string(),
});

/**
 * @summary List bins in a group
 */
export const ListBinsParams = zod.object({
  groupId: zod.coerce.string(),
});

export const ListBinsResponse = zod.object({
  bins: zod.array(
    zod.object({
      id: zod.string(),
      groupId: zod.string(),
      locationId: zod.string(),
      name: zod.string(),
      qrCode: zod.string().nullish(),
      createdAt: zod.coerce.date(),
      updatedAt: zod.coerce.date(),
    }),
  ),
});

/**
 * @summary Create a bin
 */
export const CreateBinParams = zod.object({
  groupId: zod.coerce.string(),
});

export const CreateBinBody = zod.object({
  name: zod.string(),
  locationId: zod.string(),
  qrCode: zod.string().nullish(),
});

/**
 * @summary Update a bin
 */
export const UpdateBinParams = zod.object({
  groupId: zod.coerce.string(),
  binId: zod.coerce.string(),
});

export const UpdateBinBody = zod.object({
  name: zod.string().optional(),
  locationId: zod.string().optional(),
  qrCode: zod.string().nullish(),
});

export const UpdateBinResponse = zod.object({
  bin: zod.object({
    id: zod.string(),
    groupId: zod.string(),
    locationId: zod.string(),
    name: zod.string(),
    qrCode: zod.string().nullish(),
    createdAt: zod.coerce.date(),
    updatedAt: zod.coerce.date(),
  }),
});

/**
 * @summary Delete a bin
 */
export const DeleteBinParams = zod.object({
  groupId: zod.coerce.string(),
  binId: zod.coerce.string(),
});

/**
 * @summary List items in a group
 */
export const ListItemsParams = zod.object({
  groupId: zod.coerce.string(),
});

export const ListItemsQueryParams = zod.object({
  q: zod.coerce.string().optional().describe("Search query"),
  binId: zod.coerce.string().optional().describe("Filter by bin"),
});

export const ListItemsResponse = zod.object({
  items: zod.array(
    zod.object({
      id: zod.string(),
      groupId: zod.string(),
      binId: zod.string(),
      name: zod.string(),
      photo: zod.string().nullish(),
      qrCode: zod.string().nullish(),
      createdAt: zod.coerce.date(),
      updatedAt: zod.coerce.date(),
    }),
  ),
});

/**
 * @summary Create an item
 */
export const CreateItemParams = zod.object({
  groupId: zod.coerce.string(),
});

export const CreateItemBody = zod.object({
  name: zod.string(),
  binId: zod.string(),
  photo: zod.string().nullish(),
  qrCode: zod.string().nullish(),
});

/**
 * @summary Update an item
 */
export const UpdateItemParams = zod.object({
  groupId: zod.coerce.string(),
  itemId: zod.coerce.string(),
});

export const UpdateItemBody = zod.object({
  name: zod.string().optional(),
  binId: zod.string().optional(),
  photo: zod.string().nullish(),
  qrCode: zod.string().nullish(),
});

export const UpdateItemResponse = zod.object({
  item: zod.object({
    id: zod.string(),
    groupId: zod.string(),
    binId: zod.string(),
    name: zod.string(),
    photo: zod.string().nullish(),
    qrCode: zod.string().nullish(),
    createdAt: zod.coerce.date(),
    updatedAt: zod.coerce.date(),
  }),
});

/**
 * @summary Delete an item
 */
export const DeleteItemParams = zod.object({
  groupId: zod.coerce.string(),
  itemId: zod.coerce.string(),
});

/**
 * @summary Request a presigned URL for file upload
 */
export const RequestUploadUrlBody = zod.object({
  name: zod.string(),
  size: zod.number(),
  contentType: zod.string(),
});

export const RequestUploadUrlResponse = zod.object({
  uploadURL: zod.string(),
  objectPath: zod.string(),
  metadata: zod.object({
    name: zod.string(),
    size: zod.number(),
    contentType: zod.string(),
  }),
});

/**
 * @summary Look up what a QR code points to
 */
export const LookupQrCodeParams = zod.object({
  qrCode: zod.coerce.string(),
});

export const LookupQrCodeResponse = zod.object({
  type: zod.enum(["bin", "location", "item"]),
  id: zod.string(),
  groupId: zod.string(),
});